load contacntList instead contacts with account
